+++ lib/src/components/display/number_ticker.dart
@@ -311,3 +311,177 @@ class NumberTicker extends StatelessWidget {
     );
   }
 }
+
+class FlipperCharset {
+  static const FlipperCharset numbers = FlipperCharset('0123456789');
+  static const FlipperCharset uppercase =
+      FlipperCharset('ABCDEFGHIJKLMNOPQRSTUVWXYZ');
+  static const FlipperCharset lowercase =
+      FlipperCharset('abcdefghijklmnopqrstuvwxyz');
+  static const FlipperCharset letters =
+      FlipperCharset('ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz');
+  static const FlipperCharset alphanumeric = FlipperCharset(
+      'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789');
+  static const FlipperCharset symbols =
+      FlipperCharset('!@#\$%^&*()-_=+[]{}|;:\'",.<>?/`~');
+  static const FlipperCharset all = FlipperCharset(
+      'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789!@#\$%^&*()-_=+[]{}|;:\'",.<>?/`~');
+  final String characters;
+  const FlipperCharset(this.characters);
+}
+
+class FlipperCharacter extends StatelessWidget {
+  final FlipperCharset charset;
+  final String character;
+  final Duration duration;
+  final Curve curve;
+
+  const FlipperCharacter({
+    super.key,
+    required this.charset,
+    required this.character,
+    required this.duration,
+    required this.curve,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    int index = charset.characters.indexOf(character);
+    return AnimatedValueBuilder(
+      value: index.toDouble(),
+      duration: duration,
+      curve: curve,
+      builder: (context, value, child) {
+        final nextChar = value.ceil();
+        final prevChar = value.floor();
+        if (nextChar == prevChar) {
+          final charString = nextChar == -1
+              ? ''
+              : charset.characters[nextChar % charset.characters.length];
+          return Text(charString);
+        }
+        final t = value - prevChar;
+        final nextCharString = nextChar == -1
+            ? ''
+            : charset.characters[nextChar % charset.characters.length];
+        final prevCharString = prevChar == -1
+            ? ''
+            : charset.characters[prevChar % charset.characters.length];
+        final nextOffset = Offset(0, -(1 - t));
+        final prevOffset = Offset(0, t);
+        final nextOpacity = t;
+        final prevOpacity = 1 - t;
+        Widget child = Stack(
+          children: [
+            Visibility(
+              visible: false,
+              maintainSize: true,
+              maintainAnimation: true,
+              maintainState: true,
+              child: Text(nextCharString),
+            ),
+            Positioned.fill(
+              child: Opacity(
+                opacity: prevOpacity,
+                child: LayoutBuilder(
+                  builder: (context, constraints) {
+                    return Transform.translate(
+                      offset: Offset(
+                        prevOffset.dx * constraints.maxWidth,
+                        prevOffset.dy * constraints.maxHeight,
+                      ),
+                      child: Text(prevCharString),
+                    );
+                  },
+                ),
+              ),
+            ),
+            Positioned.fill(
+              child: Opacity(
+                opacity: nextOpacity,
+                child: LayoutBuilder(
+                  builder: (context, constraints) {
+                    return Transform.translate(
+                      offset: Offset(
+                        nextOffset.dx * constraints.maxWidth,
+                        nextOffset.dy * constraints.maxHeight,
+                      ),
+                      child: Text(nextCharString),
+                    );
+                  },
+                ),
+              ),
+            ),
+          ],
+        );
+        final halfT = (t - 0.5).abs().clamp(0.0, 0.5);
+        final gradientHeight = 0.5 - halfT; // or (0.5 - halfT).clamp(0.0, 0.5)
+        return AnimatedSize(
+          duration: duration,
+          curve: curve,
+          child: _FlipperGradientMask(
+            gradientHeight: gradientHeight,
+            child: child,
+          ),
+        );
+      },
+    );
+  }
+}
+
+class _FlipperGradientMask extends StatelessWidget {
+  final Widget child;
+  final double gradientHeight; // fraction
+  const _FlipperGradientMask({
+    required this.child,
+    required this.gradientHeight,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return ShaderMask(
+      shaderCallback: (bounds) {
+        return LinearGradient(
+          begin: Alignment.topCenter,
+          end: Alignment.bottomCenter,
+          colors: [
+            Colors.white.withAlpha(0),
+            Colors.white,
+            Colors.white,
+            Colors.white.withAlpha(0),
+          ],
+          stops: [0.0, gradientHeight, 1 - gradientHeight, 1.0],
+        ).createShader(bounds);
+      },
+      child: child,
+    );
+  }
+}
+
+class TextFlipper extends StatelessWidget {
+  final FlipperCharset charset;
+  final String text;
+  final Duration duration;
+  final Curve curve;
+  const TextFlipper({
+    super.key,
+    this.charset = FlipperCharset.all,
+    required this.text,
+    this.duration = const Duration(milliseconds: 500),
+    this.curve = Curves.easeInOut,
+  });
+  @override
+  Widget build(BuildContext context) {
+    return Row(
+      mainAxisSize: MainAxisSize.min,
+      children: text.characters
+          .map((char) => FlipperCharacter(
+                charset: charset,
+                character: char,
+                duration: duration,
+                curve: curve,
+              ))
+          .toList(),
+    );
+  }
+}