+++ lib/src/components/layout/flex.dart
@@ -1,4 +1,4 @@
-import 'dart:ui' show Offset;
+import 'dart:ui' show Offset, Clip;
 
 import 'package:flutter/foundation.dart' as foundation;
 import 'package:flutter/rendering.dart' as rendering;
@@ -177,8 +177,25 @@ mixin PaintOrderMixin on rendering.RenderBox {
     return head;
   }
 
+  Clip? get optionalClipBehavior => null;
+
   /// Paints children in sorted paint order.
   void paintSorted(rendering.PaintingContext context, Offset offset) {
+    final clipBehavior = optionalClipBehavior;
+    if (clipBehavior != null && clipBehavior != Clip.none) {
+      context.pushClipRect(
+        needsCompositing,
+        offset,
+        Offset.zero & size,
+        _paintSorted,
+        clipBehavior: clipBehavior,
+      );
+    } else {
+      _paintSorted(context, offset);
+    }
+  }
+
+  void _paintSorted(rendering.PaintingContext context, Offset offset) {
     rendering.RenderBox? child = firstSortedChild;
     while (child != null) {
       final parentData = child.parentData! as PaintOrderParentDataMixin;
@@ -240,6 +257,9 @@ class RenderFlex extends rendering.RenderFlex with PaintOrderMixin {
 
   @override
   rendering.RenderBox? get paintOrderFirstChild => firstChild;
+
+  @override
+  Clip? get optionalClipBehavior => clipBehavior;
 
   @override
   void setupParentData(rendering.RenderBox child) {
@@ -402,6 +422,7 @@ class Row extends Flex {
     super.textBaseline,
     super.spacing,
     super.children,
+    super.clipBehavior,
   }) : super(direction: widgets.Axis.horizontal);
 }
 
@@ -438,6 +459,7 @@ class Column extends Flex {
     super.textBaseline,
     super.spacing,
     super.children,
+    super.clipBehavior,
   }) : super(direction: widgets.Axis.vertical);
 }
 
@@ -612,6 +634,9 @@ class RenderStack extends rendering.RenderStack with PaintOrderMixin {
       child.parentData = StackParentData();
     }
   }
+
+  @override
+  Clip? get optionalClipBehavior => clipBehavior;
 
   @override
   void performLayout() {